+++ CHANGELOG.md
@@ -490,3 +490,4 @@ All notable changes to Project Genesis will be documented in this file.
 - [x] Implemented `is_deep_water_dancer` trait. Entities with this trait gain energy when on `deep-water` terrain.
 
 - Implemented `is_lava_dancer` trait. Entities with this trait gain energy when on `lava` terrain.
+- Implemented `is_wall_dancer` trait. Entities with this trait gain energy when on `wall` terrain.